+++ apps/restconf/restconf_methods.c
@@ -214,7 +214,7 @@ api_data_get2(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -227,7 +227,7 @@ api_data_get2(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -246,7 +246,7 @@ api_data_get2(clicon_handle h,
 #endif
     /* Check if error return  */
     if ((xe = xpath_first(xret, "//rpc-error")) != NULL){
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -277,7 +277,20 @@ api_data_get2(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
+		goto done;
+	    goto ok;
+	}
+	/* Check if not exists */
+	if (xlen == 0){
+	    /* 4.3: If a retrieval request for a data resource represents an 
+	       instance that does not exist, then an error response containing 
+	       a "404 Not Found" status-line MUST be returned by the server.  
+	       The error-tag value "invalid-value" is used in this case. */
+	    if (netconf_invalid_value_xml(&xe, "application", "Instance does not exist") < 0)
+		goto done;
+	    /* override invalid-value default 400 with 404 */
+	    if (api_return_err(h, r, xe, pretty, use_xml, 404) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -487,7 +500,7 @@ api_data_post(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -501,7 +514,7 @@ api_data_post(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -514,7 +527,7 @@ api_data_post(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -523,7 +536,7 @@ api_data_post(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -538,7 +551,7 @@ api_data_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -560,14 +573,15 @@ api_data_post(clicon_handle h,
 	    		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    		goto done;
 	    }
-    if (debug){
-	cbuf *ccc=cbuf_new();
-	if (clicon_xml2cbuf(ccc, xe, 0, 0) < 0)
-	    goto done;
-	clicon_debug(1, "%s XE:%s", __FUNCTION__, cbuf_get(ccc));
-    }
-
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+#if 0
+	    if (debug){
+		cbuf *ccc=cbuf_new();
+		if (clicon_xml2cbuf(ccc, xe, 0, 0) < 0)
+		    goto done;
+		clicon_debug(1, "%s XE:%s", __FUNCTION__, cbuf_get(ccc));
+	    }
+#endif
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -599,7 +613,7 @@ api_data_post(clicon_handle h,
     if (clicon_rpc_netconf(h, cbuf_get(cbx), &xret, NULL) < 0)
 	goto done;
     if ((xe = xpath_first(xret, "//rpc-error")) != NULL){
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -621,7 +635,7 @@ api_data_post(clicon_handle h,
 	/* log errors from discard, but ignore */
 	if ((xpath_first(xretdis, "//rpc-error")) != NULL)
 	    clicon_log(LOG_WARNING, "%s: discard-changes failed which may lead candidate in an inconsistent state", __FUNCTION__);
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0) /* Use original xe */
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0) /* Use original xe */
 	    goto done;
 	goto ok;
     }
@@ -835,7 +849,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -850,7 +864,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -863,7 +877,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -872,7 +886,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -887,7 +901,7 @@ api_data_put(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -907,7 +921,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -947,7 +961,7 @@ api_data_put(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -972,7 +986,7 @@ api_data_put(clicon_handle h,
 		    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		    goto done;
 		}
-		if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+		if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		    goto done;
 		goto ok;
 	    }
@@ -996,7 +1010,7 @@ api_data_put(clicon_handle h,
 			    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 			    goto done;
 			}
-			if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+			if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 			    goto done;
 			goto ok;
 		    }
@@ -1033,7 +1047,7 @@ api_data_put(clicon_handle h,
 	goto done;
 
     if ((xe = xpath_first(xret, "//rpc-error")) != NULL){
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1054,7 +1068,7 @@ api_data_put(clicon_handle h,
 	/* log errors from discard, but ignore */
 	if ((xpath_first(xretdis, "//rpc-error")) != NULL)
 	    clicon_log(LOG_WARNING, "%s: discard-changes failed which may lead candidate in an inconsistent state", __FUNCTION__);
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1187,7 +1201,7 @@ api_data_delete(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -1211,7 +1225,7 @@ api_data_delete(clicon_handle h,
     if (clicon_rpc_netconf(h, cbuf_get(cbx), &xret, NULL) < 0)
 	goto done;
     if ((xe = xpath_first(xret, "//rpc-error")) != NULL){
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1233,7 +1247,7 @@ api_data_delete(clicon_handle h,
 	/* log errors from discard, but ignore */
 	if ((xpath_first(xretdis, "//rpc-error")) != NULL)
 	    clicon_log(LOG_WARNING, "%s: discard-changes failed which may lead candidate in an inconsistent state", __FUNCTION__);
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1425,7 +1439,7 @@ api_operations_post_input(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto fail;
 	}
@@ -1438,7 +1452,7 @@ api_operations_post_input(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto fail;
 	}
@@ -1447,7 +1461,7 @@ api_operations_post_input(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto fail;
 	}
@@ -1480,7 +1494,7 @@ api_operations_post_input(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto fail;
     }
@@ -1554,7 +1568,7 @@ api_operations_post_output(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto fail;
     }
@@ -1591,7 +1605,7 @@ api_operations_post_output(clicon_handle h,
 		clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 		goto done;
 	    }
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto fail;
 	}
@@ -1724,7 +1738,7 @@ api_operations_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1743,7 +1757,7 @@ api_operations_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1754,7 +1768,7 @@ api_operations_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1783,7 +1797,7 @@ api_operations_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto done;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1823,7 +1837,7 @@ api_operations_post(clicon_handle h,
 	    clicon_err(OE_XML, EINVAL, "rpc-error not found (internal error)");
 	    goto ok;
 	}
-	if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 	    goto done;
 	goto ok;
     }
@@ -1852,7 +1866,7 @@ api_operations_post(clicon_handle h,
 	    goto done;
 	/* Local error: return it and quit */
 	if ((xe = xpath_first(xret, "rpc-reply/rpc-error")) != NULL){
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}
@@ -1861,7 +1875,7 @@ api_operations_post(clicon_handle h,
 	if (clicon_rpc_netconf_xml(h, xtop, &xret, NULL) < 0)
 	    goto done;
 	if ((xe = xpath_first(xret, "rpc-reply/rpc-error")) != NULL){
-	    if (api_return_err(h, r, xe, pretty, use_xml) < 0)
+	    if (api_return_err(h, r, xe, pretty, use_xml, 0) < 0)
 		goto done;
 	    goto ok;
 	}