+++ test/test_nacm_default.sh
@@ -115,8 +115,7 @@ EOF
 	;;
 	1) ret='{"ietf-restconf:errors" : {"error": {"error-type": "application","error-tag": "access-denied","error-severity": "error","error-message": "default deny"}}}'
 	   ;;
-	2) ret='null
-'
+	2) ret='{"ietf-restconf:errors" : {"error": {"rpc-error": {"error-type": "application","error-tag": "invalid-value","error-severity": "error","error-message": "Instance does not exist"}}}}'
 	;;
     esac
 
@@ -140,8 +139,7 @@ EOF
 	;;
 	1) ret='{"ietf-restconf:errors" : {"error": {"error-type": "application","error-tag": "access-denied","error-severity": "error","error-message": "default deny"}}}'
         ;;
-	2) ret='null
-'
+	2) ret='{"ietf-restconf:errors" : {"error": {"rpc-error": {"error-type": "application","error-tag": "invalid-value","error-severity": "error","error-message": "Instance does not exist"}}}}'
 	   ;;
 	3) ret='{"nacm-example:x": 42}
 '